requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
oauthlib==3.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
tweepy==4.4.0
urllib3==1.26.7
Werkzeug==2.0.2